massive changes for location

+++ resources/views/photos/index.blade.php
@@ -19,8 +19,14 @@
 	<a href="/activities/index/">Back to Tours</a>
 	@endif
 	<h1 style="font-size:1.5em;">
+	@guest
+		<span style="margin-left: 5px;">Photos ({{ count($photos) }})</span>
+	@else
+		@if (Auth::user()->user_type >= 100)
 		<a href="{{$add_link}}"><span class="glyphSliders glyphicon glyphicon-plus-sign" style="padding:5px;"></span></a>
 		<span style="margin-left: 5px;">{{$title}} Photos ({{ count($photos) }})</span>
+		@endif
+	@endguest
 	</h1>
 		<table class="table table-striped">
 			<tbody>
@@ -29,17 +35,35 @@
 				<tr>
 					<td>
 						<table>
-							<tr><td>{{ $photo->filename }}</td></tr>
+						
+						@guest
+						@else
+							@if (Auth::user()->user_type >= 100)
+								<tr><td>{{ $photo->filename }}</td></tr>
+							@endif
+						@endguest						
+						
 							<tr><td>{{ $photo->alt_text }}</td></tr>
 							<tr><td>{{ $photo->location }}</td></tr>
 							@if ($photo->main_flag === 1)
 							<tr><td style=""><span class="glyphSliders glyphicon glyphicon-picture"></span>{{ $photo->main_flag === 1 ? 'Main Photo' : '' }}</td></tr>
 							@endif
-							<tr><td style="padding-top:15px;"><a href="/photos/edit/{{$photo->id}}"><span class="glyphSliders glyphicon glyphicon-edit"></span></a></td></tr>
-							<tr><td style="padding-top:15px;"><a href="/photos/confirmdelete/{{$photo->id}}"><span class="glyphSliders  glyphicon glyphicon-trash"></span></a></td></tr>
+							
+							@guest
+							@else
+								@if (Auth::user()->user_type >= 100)
+									<tr><td style="padding-top:15px;"><a href="/photos/edit/{{$photo->id}}"><span class="glyphSliders glyphicon glyphicon-edit"></span></a></td></tr>
+									<tr><td style="padding-top:15px;"><a href="/photos/confirmdelete/{{$photo->id}}"><span class="glyphSliders  glyphicon glyphicon-trash"></span></a></td></tr>
+								@endif
+							@endguest
+							
 						</table>
 					</td>
-					<td><img title="{{ $photo->alt_text }}" src="{{$fullpath}}" style="width: 100%; max-width:500px"/></td>
+					<td>
+						<a href="/photos/view/{{$photo->id}}">
+							<img title="{{ $photo->alt_text }}" src="{{$fullpath}}" style="width: 100%; max-width:500px"/>
+						</a>
+					</td>
 				</tr>
 			@endforeach
 			</tbody>
